proxy-poc/requirements.txt
fastapi
uvicorn
pillow
transformers
numpy
opencv-python-headless
torch
scikit-learn
matplotlib
python-multipart
torchvision
supabase
python-dotenv
psycopg2-binary
